+++ assembly/src/tokens/mod.rs
@@ -54,7 +54,7 @@ impl<'a> Token<'a> {
     // --------------------------------------------------------------------------------------------
     pub const DOC_COMMENT_PREFIX: &str = "#!";
     pub const COMMENT_PREFIX: char = '#';
-    pub const EXPORT_ALIAS_DELIM: &str = "->";
+    pub const ALIAS_DELIM: &str = "->";
 
     // CONSTRUCTOR
     // --------------------------------------------------------------------------------------------
@@ -112,12 +112,23 @@ impl<'a> Token<'a> {
     // CONTROL TOKEN PARSERS / VALIDATORS
     // --------------------------------------------------------------------------------------------
 
-    pub fn parse_use(&self) -> Result<LibraryPath, ParsingError> {
+    pub fn parse_use(&self) -> Result<(LibraryPath, String), ParsingError> {
         assert_eq!(Self::USE, self.parts[0], "not a use");
         match self.num_parts() {
             0 => unreachable!(),
             1 => Err(ParsingError::missing_param(self)),
-            2 => validate_import_path(self.parts[1], self),
+            2 => {
+                if let Some((module_path, module_name)) =
+                    self.parts[1].split_once(Self::ALIAS_DELIM)
+                {
+                    validate_module_name(module_name, self)?;
+                    Ok((validate_import_path(module_path, self)?, module_name.to_string()))
+                } else {
+                    let module_path = validate_import_path(self.parts[1], self)?;
+                    let module_name = module_path.last().to_string();
+                    Ok((module_path, module_name))
+                }
+            }
             _ => Err(ParsingError::extra_param(self)),
         }
     }
@@ -171,7 +182,7 @@ impl<'a> Token<'a> {
 
                 // get the alias name if it exists else export it with the original name
                 let (ref_name, proc_name) = proc_name_with_alias
-                    .split_once(Self::EXPORT_ALIAS_DELIM)
+                    .split_once(Self::ALIAS_DELIM)
                     .unwrap_or((proc_name_with_alias, proc_name_with_alias));
 
                 // validate the procedure names
@@ -287,5 +298,21 @@ fn validate_proc_locals(locals: &str, token: &Token) -> Result<u16, ParsingError
             Ok(num_locals as u16)
         }
         Err(_) => Err(ParsingError::invalid_proc_locals(token, locals)),
+    }
+}
+
+/// A module name must comply with the following rules:
+/// - The name must be between 1 and 255 characters long.
+/// - The name must start with an ASCII letter.
+/// - The name can contain only ASCII letters, numbers, or underscores.
+fn validate_module_name(name: &str, token: &Token) -> Result<(), ParsingError> {
+    if name.is_empty()
+        || name.len() > crate::MAX_LABEL_LEN
+        || !name.chars().next().unwrap().is_ascii_alphabetic()
+        || !name.chars().all(|c| c.is_ascii_alphanumeric() || c == '_')
+    {
+        Err(ParsingError::invalid_module_name(token, name))
+    } else {
+        Ok(())
     }
 }